requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --no-emit-options --output-file=requirements.txt --resolver=backtracking requirements.in
#
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
    #   watchfiles
async-timeout==4.0.2
    # via redis
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via
    #   rq
    #   uvicorn
crontab==1.0.1
    # via
    #   -r requirements.in
    #   rq-scheduler
fastapi==0.95.1
    # via -r requirements.in
freezegun==1.2.2
    # via rq-scheduler
gunicorn==20.1.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.0
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.24.0
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
pydantic==1.10.7
    # via fastapi
python-dateutil==2.8.2
    # via
    #   freezegun
    #   rq-scheduler
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0
    # via uvicorn
redis==4.5.4
    # via rq
rq==1.13.0
    # via
    #   -r requirements.in
    #   rq-scheduler
rq-scheduler==0.13.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.26.1
    # via fastapi
typing-extensions==4.5.0
    # via pydantic
uvicorn[standard]==0.21.1
    # via -r requirements.in
uvloop==0.17.0
    # via uvicorn
watchfiles==0.19.0
    # via uvicorn
websockets==11.0.1
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
